:recycle: Integrate ManualTaskConverter into ManualTaskService
