--- src/converters/index.ts
@@ -1,2 +1,0 @@
-export * from './user_task_converter';
-export * from './manual_task_converter';
